requirements.txt
certifi==2021.5.30
click==7.1.2
Flask==1.1.2
Pillow==8.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
Werkzeug==1.0.1
wincertstore==0.2
tensorflow==2.5.1
tensorflow-cpu==2.9.1
numpy==1.23.1
flask-restful==0.3.9